mixingratio_to_percc {conversions}	R Documentation

Convert Concentrations

Description

This function can have a mixing ratio input i.e. ppb, and output a concentration that in per cubic centimeter. All concentrations should be stated. All temperatures should be in K all pressures in atm.

Usage

mixingratio_to_percc(conc, temp = 298, pressure = 1, mixingratio = "ppb")

Arguments

conc
numeric containing concentration value

temp
tmeperature in K

pressure
Pressure in atm

mixingratio
Mixing ratio units i.e. ppth, ppm, ppb, ppt, ppq

Value

A numeric value in molecules cm-3

[Package conversions version 0.0.0.9000 Index]